Extract speaker targeting calculations into TargetSpeaker
